\name{cellCellSimulate}
\docType{methods}
\alias{cellCellSimulate,SingleCellExperiment-method}
\alias{cellCellSimulate}

\title{
Parameter Simulate for scTensor
}
\description{
All parameters is saved to metadata slot of SingleCellExperiment object.
}
\usage{
cellCellSimulate(params = newCCSParams(), verbose = TRUE)
}

\arguments{
\item{params}{
A parameter object generated by newCCSParams().
}
\item{verbose}{
Whether the message is outputted or not (Default: TRUE).
}
}

\value{
A list object containing simcount, LR, celltype, and LR_CCI.
simcount is the synthetic count matrix, LR is the synthetic ligand-receptor pair list, celltype is the vector to specity the celltype of the each column of simcount, and LR_CCI is the vector to specify the CCI of the each row of LR.
}

\author{Koki Tsuyuzaki}

\examples{
showMethods("cellCellSimulate")
}

\keyword{methods}