\name{labelGermMale}
\alias{labelGermMale}
\docType{data}
\title{
The vector contains the celltype information and color scheme of GermMale
}
\description{
A vector with 852 length (cells).
}
\usage{data(labelGermMale)}
\details{
The Cluster label is downloaded from original paper page of Cell Stem Cell
(https://www.sciencedirect.com/science/article/pii/S1934590917300784)
}
\references{
Li L. and Dong J. and Yan L. and Yong J. et al. (2017)
Single-Cell RNA-Seq Analysis Maps Development of Human Germline Cells
and Gonadal Niche Interactions. \emph{Cell Stem Cell}, \bold{20(6)}: 858-873
}
\seealso{
\code{\link{GermMale}}, \code{\link{tsneGermMale}}.
}
\examples{
data(labelGermMale)
}
\keyword{datasets}