\name{scTensor-package}
\alias{scTensor-package}
\alias{scTensor}
\docType{package}
\title{
\packageTitle{scTensor}
}
\description{
\packageDescription{scTensor}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{scTensor}
\packageIndices{scTensor}
}
\author{
\packageAuthor{scTensor}

Maintainer: \packageMaintainer{scTensor}
}

\keyword{ package }
\seealso{\code{\link{GermMale}},\code{\link{labelGermMale}},
\code{\link{tsneGermMale}},\code{\link{cellCellSetting}},
\code{\link{cellCellDecomp}},\code{\link{cellCellReport}}}
\examples{
ls("package:scTensor")
}