\name{cellCellRanks}
\docType{methods}
\alias{cellCellRanks,SingleCellExperiment-method}
\alias{cellCellRanks}

\title{
Rank estimation of the CCI-tensor
}
\description{
SVD is performed in each mode.
}
\usage{
cellCellRanks(sce, centering=TRUE,
    mergeas=c("mean", "sum"), outer=c("*", "+"), comb=c("random", "all"),
    num.sampling=100, ftt=TRUE, thr1=0.9, thr2=0.9, thr3=0.9)
}

\arguments{
\item{sce}{
A object generated by instantization of SingleCellExperiment-class.
}
\item{centering}{
When the value is TRUE, input matrix is summarized as celltype-level vectors
(Default: TRUE).
}
\item{mergeas}{
When the centering is TRUE, "mean" (celltype-level mean vector)
or "sum" (celltype-level sum vector) is calculated (Default: "mean").
}
\item{outer}{
When the centering is TRUE, "*" (Kronecker product) or "+" (Kronecker sum) or is calculated (Default: "+").
}
\item{comb}{
When the centering is FALSE, "random" (random cell-cell pairing)
or "all" (all possible cell-cell pairing) is calculed (Default: "random").
}
\item{num.sampling}{
The number of random sampling used
(Default: 100).
}
\item{ftt}{
When the value is TRUE, Freeman-Tuckey transformation (FTT) of data matrix is calculated (Default: TRUE).
}
\item{thr1}{
The threshold for selection of top eigenvalues of mode-1 matricised data tensor (Default: 0.9 (0 to 1)).
}
\item{thr2}{
The threshold for selection of top eigenvalues of mode-2 matricised data tensor (Default: 0.9 (0 to 1)).
}
\item{thr3}{
The threshold for selection of top eigenvalues of mode-3 matricised data tensor (Default: 0.9 (0 to 1)).
}
}

\value{
A vector with three elements, in which each element means the estimated ranks in mode-1, 2 and 3 matricization.
}
\seealso{
\code{\link[SingleCellExperiment]{SingleCellExperiment}}.
}

\author{Koki Tsuyuzaki}

\examples{
showMethods("cellCellRanks")
}

\keyword{methods}