\name{cellCellReport}
\docType{methods}
\alias{cellCellReport,SingleCellExperiment-method}
\alias{cellCellReport}

\title{
HTML report of the result of scTensor
}
\description{
The result is saved as HTML report which contains with multiple files.
}
\usage{
cellCellReport(sce, reducedDimNames,
    out.dir=tempdir(), html.open=FALSE,
    title="The result of scTensor",
    author="The person who runs this script", assayNames = "counts", thr=100,
    top="full", p=0.05, upper=20,
    goenrich=TRUE, meshenrich=TRUE, reactomeenrich=TRUE,
    doenrich=TRUE, ncgenrich=TRUE, dgnenrich=TRUE)
}

\arguments{
\item{sce}{
A object generated by instantization of SingleCellExperiment-class.
}
\item{reducedDimNames}{
The name of two-dimentional data saved in reducedDimNames slot of
SingleCellExperiment object.
}
\item{out.dir}{
The output directory for saving HTML report (out.dir: tempdir()).
}
\item{html.open}{
Whether the result of HTML report is opened when the calculation is finished
(Default: FALSE).
}
\item{title}{
The title of HTML report (Default: "The result of scTensor").
}
\item{author}{
The author of HTML report (Default: "The person who runs this script").
}
\item{assayNames}{
The unit of gene expression for using scTensor (e.g. normcounts, cpm...etc) (Default: "counts").
}
\item{thr}{
The threshold for selection of top pecentage of core tensor elements
(Default: 100 (1 to 100)).
}
\item{top}{
top genes in each (*,*,*)-pattern which are selected and summarized in the report (Default: "full")
}
\item{p}{
The threshold of p-value of the enrichment analysis (Default: 1E-2)
}
\item{upper}{
The maxium number of HTML reports generates (Default: 20)
}
\item{goenrich}{
Whether GO-Enrichment analysis is performed (Default: TRUE)
}
\item{meshenrich}{
Whether MeSH-Enrichment analysis is performed (Default: TRUE)
}
\item{reactomeenrich}{
Whether Reactome-Enrichment analysis is performed (Default: TRUE)
}
\item{doenrich}{
Whether DO-Enrichment analysis is performed (Default: TRUE)
}
\item{ncgenrich}{
Whether NCG-Enrichment analysis is performed (Default: TRUE)
}
\item{dgnenrich}{
Whether DGN-Enrichment analysis is performed (Default: TRUE)
}
}

\value{
The result is saved as HTML report which contains with multiple files.
}
\seealso{
\code{\link[SingleCellExperiment]{SingleCellExperiment}}.
}

\author{Koki Tsuyuzaki}

\examples{
	if(interactive()){
	# Package Loading
	library("SingleCellExperiment")
	library("LRBase.Hsa.eg.db")
	library("MeSH.Hsa.eg.db")

	# Data Loading
	data(GermMale)
	data(labelGermMale)
	data(tsneGermMale)

	# SingleCellExperiment Object
	sce <- SingleCellExperiment(assays=list(counts = GermMale))
	reducedDims(sce) <- SimpleList(TSNE=tsneGermMale$Y)

	# User's Original Normalization Function
	CPMED <- function(input){
	    libsize <- colSums(input)
	    median(libsize) * t(t(input) / libsize)
	}
	# Normalization
	normcounts(sce) <- log10(CPMED(counts(sce)) + 1)
	# Registration of required information into metadata(sce)
	cellCellSetting(sce, LRBase.Hsa.eg.db, labelGermMale, names(labelGermMale))
	# Rank Estimation
	rks <- cellCellRanks(sce, assayNames="normcounts")
	# CCI Tensor Decomposition
	cellCellDecomp(sce, ranks=rks$selected, assayNames="normcounts")
	# HTML Report
	cellCellReport(sce, reducedDimNames="TSNE", assayNames="normcounts",
	    title="Cell-cell interaction within Germline_Male, GSE86146",
	    author="Koki Tsuyuzaki", html.open=TRUE, upper=2,
        goenrich=TRUE, meshenrich=FALSE, reactomeenrich=FALSE,
        doenrich=FALSE, ncgenrich=FALSE, dgnenrich=FALSE)
	if(.Device == "quartz"){
		dev.off()
	}
    }
}

\keyword{methods}