\name{cellCellReport}
\docType{methods}
\alias{cellCellReport,SingleCellExperiment-method}
\alias{cellCellReport}

\title{
HTML report of the result of scTensor
}
\description{
The result is saved as HTML report which contains with multiple files.
}
\usage{
cellCellReport(sce, reducedDimNames,
    out.dir=tempdir(), html.open=FALSE,
    title="The result of scTensor",
    author="The person who runs this script", thr=40, top="full", p=0.05, upper=100)
}

\arguments{
\item{sce}{
A object generated by instantization of SingleCellExperiment-class.
}
\item{reducedDimNames}{
The name of two-dimentional data saved in reducedDimNames slot of
SingleCellExperiment object.
}
\item{out.dir}{
The output directory for saving HTML report (out.dir: tempdir()).
}
\item{html.open}{
Whether the result of HTML report is opened when the calculation is finished
(Default: FALSE).
}
\item{title}{
The title of HTML report (Default: "The result of scTensor").
}
\item{author}{
The author of HTML report (Default: "The person who runs this script").
}
\item{thr}{
The threshold for selection of top pecentage of core tensor elements
(Default: 40 (1 to 100)).
}
\item{top}{
top genes in each (*,*,*)-pattern which are selected and summarized in the report (Default: "full")
}
\item{p}{
The threshold of p-value of the enrichment analysis (Default: 1E-2)
}
\item{upper}{
The maxium number of HTML reports generates (Default: 100)
}
}

\value{
The result is saved as HTML report which contains with multiple files.
}
\seealso{
\code{\link[SingleCellExperiment]{SingleCellExperiment}}.
}

\author{Koki Tsuyuzaki}

\examples{
	if(interactive()){
	library("SingleCellExperiment")
	library("LRBase.Hsa.eg.db")
	library("MeSH.Hsa.eg.db")

	data(GermMale)
	data(labelGermMale)
	data(tsneGermMale)
	sce <- SingleCellExperiment(assays=list(counts = GermMale))
	reducedDims(sce) <- SimpleList(TSNE=tsneGermMale$Y)
	cellCellSetting(sce, LRBase.Hsa.eg.db, labelGermMale, names(labelGermMale))
	rks <- cellCellRanks(sce)
	cellCellDecomp(sce, ranks=rks$selected)
	cellCellReport(sce, reducedDimNames="TSNE",
	    title="Cell-cell interaction within Germline_Male, GSE86146",
	    author="Koki Tsuyuzaki", thr=20, html.open=TRUE)
	try(dev.off())
    }
}

\keyword{methods}