\name{cellCellReport}
\docType{methods}
\alias{cellCellReport,SingleCellExperiment-method}
\alias{cellCellReport}

\title{
HTML report of the result of scTensor
}
\description{
The result is saved as HTML report which contains with multiple files.
}
\usage{
cellCellReport(sce, reducedDimNames,
    out.dir=NULL, html.open=FALSE,
    title="The result of scTensor",
    author="The person who runs this script", thr=80, top="full", cl=NULL)
}

\arguments{
\item{sce}{
A object generated by instantization of SingleCellExperiment-class.
}
\item{reducedDimNames}{
The name of two-dimentional data saved in reducedDimNames slot of
SingleCellExperiment object.
}
\item{out.dir}{
The output directory for saving HTML report (out.dir: NULL).
}
\item{html.open}{
Whether the result of HTML report is opened when the calculation is finished
(Default: FALSE).
}
\item{title}{
The title of HTML report (Default: "The result of scTensor").
}
\item{author}{
The author of HTML report (Default: "The person who runs this script").
}
\item{thr}{
The threshold for selection of top pecentage of core tensor elements
(Default: 80 (1 to 100)).
}
\item{top}{
top genes in each (*,*,*)-pattern which are selected and summarized in the report (Default: "full")
}
\item{cl}{
Cluster set by snow package. If the cl is set, report is calculated in parallel (Default: NULL)
}
}

\value{
The result is saved as HTML report which contains with multiple files.
}
\seealso{
\code{\link[SingleCellExperiment]{SingleCellExperiment}}.
}

\author{Koki Tsuyuzaki}

\examples{
showMethods("cellCellReport")
}

\keyword{methods}